+++ src/components/profile/ProfileButton.tsx
@@ -1,0 +1,28 @@
+import { useAuthStore, useGithubAuthStore } from "../../store/auth";
+import { Link } from "react-router-dom";
+
+const ProfileButton = () => {
+  const githubAuthStore = useGithubAuthStore();
+  return (
+    <section className="flex flex-col w-full h-full gap-4">
+      {githubAuthStore.isLogin && (
+        <>
+          <Link
+            to="/addproject"
+            className="btn hover:btn-ghost btn-outline w-full"
+          >
+            프로젝트 등록
+          </Link>
+          <button className="btn hover:btn-ghost btn-outline w-full">
+            내 프로젝트 보기
+          </button>
+        </>
+      )}
+      <button className="btn hover:btn-ghost btn-outline w-full">
+        좋아요 한 프로젝트 보기
+      </button>
+    </section>
+  );
+};
+
+export default ProfileButton;
