+++ src/constants/techStacks.ts
@@ -1,13 +1,13 @@
 export const TECHSTACKS = [
-  { title: "React", bgColor: "#61dafb", txtColor: "black" },
-  { title: "Python", bgColor: "#3676AB", txtColor: "white" },
-  { title: "JavaScript", bgColor: "#f7df1e", txtColor: "black" },
+  { title: "react", bgColor: "#61dafb", txtColor: "black" },
+  { title: "python", bgColor: "#3676AB", txtColor: "white" },
+  { title: "js", bgColor: "#f7df1e", txtColor: "black" },
   { title: "html", bgColor: "#e34f26", txtColor: "white" },
-  { title: "Css", bgColor: "#264de4", txtColor: "white" },
+  { title: "css", bgColor: "#264de4", txtColor: "white" },
   { title: "node.js", bgColor: "#339933", txtColor: "white" },
   { title: "java", bgColor: "#007396", txtColor: "white" },
   { title: "ruby", bgColor: "#cc342d", txtColor: "white" },
   { title: "angular", bgColor: "#dd0031", txtColor: "white" },
-  { title: "Vue.js", bgColor: "#4fc08d", txtColor: "white" },
+  { title: "vue", bgColor: "#4fc08d", txtColor: "white" },
   { title: "c#", bgColor: "#68217a", txtColor: "white" },
 ];
