+++ src/pages/ProfilePage.tsx
@@ -1,10 +1,10 @@
 import { useAuthStore, useGithubAuthStore } from "../store/auth";
 import { Navigate, Link } from "react-router-dom";
 import useLogout from "../hooks/useLogout";
-import Input from "../components/ui/Input";
 import { useState } from "react";
 import Header from "../components/ui/Header";
-import { GTIHUB_OAUTH_URL } from "../constants/authUrl";
+import UserProfile from "../components/profile/UserProfile";
+import ProfileButton from "../components/profile/ProfileButton";
 
 const ProfilePage = () => {
   const authStore = useAuthStore();
@@ -17,58 +17,15 @@ const ProfilePage = () => {
     return <Navigate to="/"></Navigate>;
   }
 
-  const githubOAuthHandler = () => {
-    window.location.assign(GTIHUB_OAUTH_URL);
-  };
-
   return (
     <div className="flex flex-col items-center w-screen min-h-screen">
       <Header />
 
-      <section className="flex grow flex-col w-full h-full items-center justify-center  gap-4">
-        <form id="loginForm" className="flex flex-col items-center gap-4 w-80">
-          <Input
-            defaultValue={authStore.userName}
-            label="아이디"
-            type="text"
-            required
-          />
-          {githubAuthStore.isLogin ? (
-            <div className="flex flex-col w-full">
-              <label htmlFor="github" className="label">
-                <span className="label-text">깃허브 계정</span>
-              </label>
-              <div className="flex items-center gap-4">
-                <div className="avatar online">
-                  <div className="w-12 rounded-full">
-                    <img src={githubAuthStore.userImage} />
-                  </div>
-                </div>
-                <p>{githubAuthStore.userName}</p>
-              </div>
-            </div>
-          ) : (
-            <div onClick={githubOAuthHandler} className="btn btn-success w-80">
-              Github 아이디 연동하기
-            </div>
-          )}
-        </form>
-        {githubAuthStore.isLogin && (
-          <>
-            <Link
-              to="/addproject"
-              className="btn hover:btn-ghost btn-outline w-80"
-            >
-              프로젝트 등록
-            </Link>
-            <button className="btn hover:btn-ghost btn-outline w-80">
-              내 프로젝트 보기
-            </button>
-          </>
-        )}
-        <button className="btn hover:btn-ghost btn-outline w-80">
-          좋아요 한 프로젝트 보기
-        </button>
+      <section className="flex grow flex-col w-80 items-center justify-center gap-4">
+        {/* 유저 정보 */}
+        <UserProfile />
+        {/* 유저 관련 버튼 */}
+        <ProfileButton />
       </section>
     </div>
   );
