+++ tsconfig.json
@@ -19,6 +19,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
+  "include": ["src", "src/constants/authUrl.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
